refactor(eslint): Extract restricted-path zones to single source of truth
Move both `architecturalZones` and `routeIsolationZones` to
`development/eslint-restricted-paths-zones.js` so the same definitions
power the base rule (`.eslintrc.base.js`) AND the router-registry
override (`.eslintrc.js`). Eliminates the manual duplication flagged
in PR review and removes the silent-drift risk: any change to the
zones is now made once and applies everywhere.

The file lives outside `.eslintrc*.js` because ESLint v8 validates
every config it loads and rejects unknown top-level exports, which
ruled out exporting the constants from `.eslintrc.base.js` directly.

Refs: WPC-402, cursor[bot] review on PR #42735
